+++ ios/escape/escape/Profile/ProfileView.swift
@@ -9,33 +9,31 @@ import Supabase
 import SwiftUI
 
 struct ProfileView: View {
-    @State var username = ""
-    @State var fullName = ""
-    @State var website = ""
-
-    @State var isLoading = false
+    @State private var controller = ProfileController()
 
     var body: some View {
         NavigationStack {
             Form {
                 Section {
-                    TextField("Username", text: $username)
+                    TextField("Username", text: $controller.username)
                         .textContentType(.username)
                         .textInputAutocapitalization(.never)
-                    TextField("Full name", text: $fullName)
+                    TextField("Full name", text: $controller.fullName)
                         .textContentType(.name)
-                    TextField("Website", text: $website)
+                    TextField("Website", text: $controller.website)
                         .textContentType(.URL)
                         .textInputAutocapitalization(.never)
                 }
 
                 Section {
                     Button("Update profile") {
-                        updateProfileButtonTapped()
+                        Task {
+                            await controller.updateProfile()
+                        }
                     }
                     .bold()
 
-                    if isLoading {
+                    if controller.isLoading {
                         ProgressView()
                     }
                 }
@@ -45,60 +43,14 @@ struct ProfileView: View {
                 ToolbarItem(placement: .topBarLeading) {
                     Button("Sign out", role: .destructive) {
                         Task {
-                            try? await supabase.auth.signOut()
+                            try? await controller.signOut()
                         }
                     }
                 }
             })
         }
         .task {
-            await getInitialProfile()
-        }
-    }
-
-    func getInitialProfile() async {
-        do {
-            let currentUser = try await supabase.auth.session.user
-
-            let profile: Profile =
-                try await supabase
-                    .from("profiles")
-                    .select()
-                    .eq("id", value: currentUser.id)
-                    .single()
-                    .execute()
-                    .value
-
-            username = profile.username ?? ""
-            fullName = profile.fullName ?? ""
-            website = profile.website ?? ""
-
-        } catch {
-            debugPrint(error)
-        }
-    }
-
-    func updateProfileButtonTapped() {
-        Task {
-            isLoading = true
-            defer { isLoading = false }
-            do {
-                let currentUser = try await supabase.auth.session.user
-
-                try await supabase
-                    .from("profiles")
-                    .update(
-                        UpdateProfileParams(
-                            username: username,
-                            fullName: fullName,
-                            website: website
-                        )
-                    )
-                    .eq("id", value: currentUser.id)
-                    .execute()
-            } catch {
-                debugPrint(error)
-            }
+            await controller.getInitialProfile()
         }
     }
 }
